Generating phylogenetic forests

A mutation engine can label every node of a samples forest by mutations and produce a consistent phylogenetic forest.

Details

The mutations are randomly generated according to three factors:

the mutational rates of the species involved in the samples forest

the genotypical specification of the mutants involved in the sample forest, i.e., the somatic mutations characterising the mutant genotypes

the SBS and ID signatures active along the species simulation

The data of points 1 and 2 are provided to the mutation engine by the method MutationEngine$add_mutant(). Instead, the active signatures are defined by using the method MutationEngine$add_exposure().

The initialisation of a MutationEngine object requires a reference sequence and the SBS and ID mutational signatures. An SBS index and a ID index of the reference sequence are then automatically built. This process may take time depending on the size of the reference sequence. Hence, the downloaded files together with the built indices are saved on the disk for subsequent MutationEngine constructions.

The objects of this class are generated by the function MutationEngine() and provide the following methods and properties:

add_exposure() adds exposures to the mutation engine.

add_mutant() adds a mutant genomic specification to the mutation engine.

change_rates_from() changes the passenger rates from a specified time.

get_SNV_signatures() returns the SNV signatures.

get_active_germline() returns the active germline subject.

get_genome_info() returns a data frame reporting reference genome information.

get_germline_subjects() returns the available germline subjects.

get_indel_signatures() returns the indel signatures.

get_known_drivers() returns the data frame of the known driver mutations.

get_population_descriptions() returns the available population descriptions.

get_species_info() returns the registered species and their rates.

infinite_sites_model is a Boolean property to enable/disable the infinite sites model.

place_mutations() places mutations on a SampleForest object and generates a PhylogeneticForest object.

set_germline_subject() sets the active germline subject.

See also