Getting the genome bulk allelic fragmentation

This method returns a data frame representing the bulk allelic fragmentation of the genome.

Arguments

sample_name

The name of the sample whose bulk allelic fragmentation is aimed (optional).

Value

A data frame reporting, for each genomic fragment and for all the allelic type on the genomic fragment, the chromosome (chr), the first base position (begin), the last base position (end), the number of copy of the major and minor alleles (major and minor, respectively), and the ratio between the number of cells exhibiting this allelic type and the total number of cells in the sample.

See also

Examples

# use a phylogenetic forest example
forest <- example("PhylogeneticForest")

# get the genome bulk allelic fragmentation
forest$get_bulk_allelic_fragmentation()
#>    chr    begin      end major minor        ratio
#> 1   22        1  5009999     2     2 1.0000000000
#> 2   22  5010000  5209999     2     0 1.0000000000
#> 3   22  5210000 10303469     2     2 1.0000000000
#> 4   22 10303470 10503469     4     2 1.0000000000
#> 5   22 10503470 17495000     2     2 1.0000000000
#> 6   22 17495001 18551894     2     2 0.9992977528
#> 7   22 17495001 18551894     3     2 0.0007022472
#> 8   22 18551895 19301344     2     1 0.0168539326
#> 9   22 18551895 19301344     2     2 0.9824438202
#> 10  22 18551895 19301344     3     2 0.0007022472
#> 11  22 19301345 20778475     2     1 0.0168539326
#> 12  22 19301345 20778475     2     2 0.9831460674
#> 13  22 20778476 23889288     2     2 1.0000000000
#> 14  22 23889289 24394000     2     2 0.9992977528
#> 15  22 23889289 24394000     3     2 0.0007022472
#> 16  22 24394001 27742499     2     1 0.0028089888
#> 17  22 24394001 27742499     2     2 0.9964887640
#> 18  22 24394001 27742499     3     2 0.0007022472
#> 19  22 27742500 29065001     2     1 0.0028089888
#> 20  22 27742500 29065001     2     2 0.9950842697
#> 21  22 27742500 29065001     3     2 0.0021067416
#> 22  22 29065002 29065252     2     2 0.9978932584
#> 23  22 29065002 29065252     3     2 0.0021067416
#> 24  22 29065253 29066121     2     2 0.9985955056
#> 25  22 29065253 29066121     3     2 0.0014044944
#> 26  22 29066122 30007695     2     1 0.0028089888
#> 27  22 29066122 30007695     2     2 0.9957865169
#> 28  22 29066122 30007695     3     2 0.0014044944
#> 29  22 30007696 31096964     2     1 0.0028089888
#> 30  22 30007696 31096964     2     2 0.9950842697
#> 31  22 30007696 31096964     3     2 0.0021067416
#> 32  22 31096965 33371607     2     1 0.0028089888
#> 33  22 31096965 33371607     2     2 0.9922752809
#> 34  22 31096965 33371607     3     2 0.0049157303
#> 35  22 33371608 36855070     2     2 0.9950842697
#> 36  22 33371608 36855070     3     2 0.0049157303
#> 37  22 36855071 38608409     2     2 0.9957865169
#> 38  22 36855071 38608409     3     2 0.0042134831
#> 39  22 38608410 39617000     2     2 0.9922752809
#> 40  22 38608410 39617000     3     2 0.0077247191
#> 41  22 39617001 40287242     2     1 0.0070224719
#> 42  22 39617001 40287242     2     2 0.9852528090
#> 43  22 39617001 40287242     3     2 0.0077247191
#> 44  22 40287243 41772801     2     1 0.0070224719
#> 45  22 40287243 41772801     2     2 0.9887640449
#> 46  22 40287243 41772801     3     2 0.0042134831
#> 47  22 41772802 42378557     2     2 0.9957865169
#> 48  22 41772802 42378557     3     2 0.0042134831
#> 49  22 42378558 46434889     2     2 0.9957865169
#> 50  22 42378558 46434889     3     2 0.0042134831
#> 51  22 46434890 48092714     2     2 0.9985955056
#> 52  22 46434890 48092714     3     2 0.0014044944
#> 53  22 48092715 48837426     2     2 0.9971910112
#> 54  22 48092715 48837426     3     2 0.0028089888
#> 55  22 48837427 49296876     2     1 0.0196629213
#> 56  22 48837427 49296876     2     2 0.9775280899
#> 57  22 48837427 49296876     3     2 0.0028089888
#> 58  22 49296877 49687500     2     2 0.8834269663
#> 59  22 49296877 49687500     3     1 0.0196629213
#> 60  22 49296877 49687500     3     2 0.0955056180
#> 61  22 49296877 49687500     3     3 0.0014044944
#> 62  22 49687501 50429352     2     2 0.8848314607
#> 63  22 49687501 50429352     3     1 0.0196629213
#> 64  22 49687501 50429352     3     2 0.0955056180
#> 65  22 50429353 50827293     2     2 0.8848314607
#> 66  22 50429353 50827293     3     2 0.1151685393
#> 67  22 50827294 51304566     2     2 1.0000000000