An iterator class over phylogenetic forest nodes

Iterators over phylogenetic forest nodes.

Details

This class represents iterators over phylogenetic forest nodes. The objects of this class are built by get_node_tour() and provide the following methods and properties:

node is an object of the class SampleForestNode and represents the node pointed by the iterator.

label represents the label of the of the node pointed by the iterator. The presence of this field depends on the type of the get_node_tour()'s parameters used to create the tour object (OPTIONAL).

genome is an object of the class GenomeMutations that represent the genome of the node pointed by the iterator (OPTIONAL).

step() moves the iterator to the next node in the tour.

done is a Boolean flag that is set to TRUE only when the tour ended.

See also