Getting snapshot information

This method returns the data frame of the snapshots

Value

A data frame consisting of four columns: time, clock, cells, and file. Each row represents a snapshot. The column time stores the snapshot time. The columns clock and cells maintain the simulated time and the number of tumour cells at the snapshot time, respectively. Finally, the column file contains the snapshot file path.

See also

Examples

# set the random seed
set.seed(0)

# build a simulation
sim <- TissueSimulation()

# add a mutant
sim$add_mutant("A", list(duplication = 3, death = 1))

# place a cell
sim$place_cell("A", 500, 500)

# get snapshot information
sim$get_snapshot_info()
#>                  time clock cells
#> 1 2026-08-19 13:50:59     0     1
#>                                                                                                   file
#> 1 /var/folders/tb/jqmdpgxs2t5129bny6pb96680000gn/T/ProCESS_alberto_6dde97b1/snapshot_65965045594bd.dat

# take a new snapshot every 10 simulated time units
sim$snapshot_triggers <- list("clock interval" = 10)

# let the simulation evolve up to time 30
sim$run_up_to_time(30)
#> 
 [██████████████--------------------------] 33% [00m:00s] Saving snapshot                                                                     

 [███████████████████---------------------] 45% [00m:00s] Cells: 16433                                                                        

 [███████████████████████-----------------] 55% [00m:01s] Cells: 26310                                                                        

 [█████████████████████████---------------] 61% [00m:02s] Cells: 34567                                                                        

 [███████████████████████████-------------] 66% [00m:03s] Saving snapshot                                                                     

 [███████████████████████████-------------] 67% [00m:03s] Cells: 41915                                                                        

 [█████████████████████████████-----------] 71% [00m:04s] Cells: 48775                                                                        

 [███████████████████████████████---------] 75% [00m:05s] Cells: 54640                                                                        

 [████████████████████████████████--------] 79% [00m:06s] Cells: 60260                                                                        

 [█████████████████████████████████-------] 82% [00m:07s] Cells: 65860                                                                        

 [██████████████████████████████████------] 84% [00m:08s] Cells: 70558                                                                        

 [███████████████████████████████████-----] 87% [00m:09s] Cells: 75066                                                                        

 [████████████████████████████████████----] 89% [00m:10s] Cells: 79587                                                                        

 [█████████████████████████████████████---] 91% [00m:11s] Cells: 82910                                                                        

 [██████████████████████████████████████--] 93% [00m:12s] Cells: 87055                                                                        

 [███████████████████████████████████████-] 95% [00m:13s] Cells: 90899                                                                        

 [███████████████████████████████████████-] 96% [00m:14s] Cells: 94492                                                                        

 [████████████████████████████████████████] 98% [00m:15s] Cells: 98173                                                                        

 [████████████████████████████████████████] 100% [00m:16s] Saving snapshot                                                                    

# get snapshot information
sim$get_snapshot_info()
#>                  time    clock  cells
#> 1 2026-08-19 13:50:59  0.00000      1
#> 2 2026-08-19 13:50:59 10.00003   7879
#> 3 2026-08-19 13:51:03 20.00003  41040
#> 4 2026-08-19 13:51:16 30.00000 101378
#>                                                                                                   file
#> 1 /var/folders/tb/jqmdpgxs2t5129bny6pb96680000gn/T/ProCESS_alberto_6dde97b1/snapshot_65965045594bd.dat
#> 2 /var/folders/tb/jqmdpgxs2t5129bny6pb96680000gn/T/ProCESS_alberto_6dde97b1/snapshot_65965045a6b3b.dat
#> 3 /var/folders/tb/jqmdpgxs2t5129bny6pb96680000gn/T/ProCESS_alberto_6dde97b1/snapshot_659650490b5fd.dat
#> 4 /var/folders/tb/jqmdpgxs2t5129bny6pb96680000gn/T/ProCESS_alberto_6dde97b1/snapshot_659650558086a.dat