Searching for a rectangular tissue sample

This method searches a rectangular tissue sample.

Arguments

min_num_of_cells

A named integer vector reporting the minimum number of cells per species or mutant.

num_of_cells

The number of cells in the searched sample.

width

The width of the searched sample.

height

The height of the searched sample.

Value

If a rectangular sample satisfying the provided constraints can be found, the corresponding rectangle.

Details

The aimed sample must satisfy the specified number of cells. The size of the sample is also provided a parameter of the method. The complexity of this method is \(O(|\textrm{tissue width}|*|\textrm{tissue height}|)\).

See also

Examples

# set the seed of the random number generator
set.seed(0)

# create a simulation
sim <- TissueSimulation()
sim$death_activation_level <- 50
sim$add_mutant(name = "A", growth_rate = 0.2, death_rate = 0.01)
sim$place_cell("A", 500, 500)
sim$run_up_to_size(species = "A", num_of_cells = 500)
#> 
 [████████████████████████████████████████] 100% [00m:00s] Saving snapshot                         

sim$add_mutant(name = "B", growth_rate = 0.3, death_rate = 0.01)
sim$mutate_progeny(sim$choose_cell_in("A"), "B")
sim$run_up_to_size(species = "B", num_of_cells = 1000)
#> 
 [██--------------------------------------] 4% [00m:00s] Cells: 46904                              

 [███-------------------------------------] 6% [00m:01s] Cells: 75801                              

 [████------------------------------------] 8% [00m:02s] Cells: 96918                              

 [████------------------------------------] 8% [00m:03s] Cells: 116538                             

 [████------------------------------------] 9% [00m:04s] Cells: 132044                             

 [████------------------------------------] 9% [00m:05s] Cells: 146384                             

 [████------------------------------------] 9% [00m:05s] Cells: 159675                             

 [████------------------------------------] 9% [00m:06s] Cells: 171746                             

 [█████-----------------------------------] 10% [00m:08s] Cells: 183240                            

 [█████-----------------------------------] 10% [00m:09s] Cells: 194131                            

 [█████-----------------------------------] 11% [00m:10s] Cells: 202865                            

 [█████-----------------------------------] 10% [00m:11s] Cells: 212226                            

 [█████-----------------------------------] 10% [00m:12s] Cells: 221586                            

 [█████-----------------------------------] 11% [00m:13s] Cells: 230719                            

 [█████-----------------------------------] 11% [00m:14s] Cells: 239434                            

 [█████-----------------------------------] 12% [00m:15s] Cells: 247327                            

 [█████-----------------------------------] 12% [00m:16s] Cells: 254934                            

 [██████----------------------------------] 13% [00m:17s] Cells: 261772                            

 [██████----------------------------------] 13% [00m:18s] Cells: 268902                            

 [█████-----------------------------------] 12% [00m:19s] Cells: 276051                            

 [█████-----------------------------------] 12% [00m:20s] Cells: 283257                            

 [██████----------------------------------] 13% [00m:21s] Cells: 289873                            

 [██████----------------------------------] 14% [00m:22s] Cells: 296611                            

 [██████----------------------------------] 13% [00m:23s] Cells: 303231                            

 [██████----------------------------------] 14% [00m:24s] Cells: 309783                            

 [██████----------------------------------] 14% [00m:25s] Cells: 316341                            

 [██████----------------------------------] 14% [00m:26s] Cells: 322639                            

 [██████----------------------------------] 13% [00m:27s] Cells: 328552                            

 [██████----------------------------------] 14% [00m:28s] Cells: 334365                            

 [██████----------------------------------] 14% [00m:29s] Cells: 340307                            

 [███████---------------------------------] 15% [00m:30s] Cells: 346296                            

 [███████---------------------------------] 15% [00m:31s] Cells: 351939                            

 [██████----------------------------------] 14% [00m:32s] Cells: 357834                            

 [███████---------------------------------] 15% [00m:33s] Cells: 363467                            

 [███████---------------------------------] 15% [00m:34s] Cells: 369028                            

 [███████---------------------------------] 17% [00m:35s] Cells: 374699                            

 [███████---------------------------------] 17% [00m:36s] Cells: 380293                            

 [████████--------------------------------] 18% [00m:37s] Cells: 385679                            

 [████████--------------------------------] 18% [00m:38s] Cells: 390824                            

 [████████--------------------------------] 18% [00m:39s] Cells: 396155                            

 [████████--------------------------------] 19% [00m:40s] Cells: 401447                            

 [████████--------------------------------] 19% [00m:41s] Cells: 406840                            

 [████████--------------------------------] 19% [00m:42s] Cells: 412253                            

 [████████--------------------------------] 19% [00m:43s] Cells: 417430                            

 [█████████-------------------------------] 20% [00m:44s] Cells: 422344                            

 [█████████-------------------------------] 20% [00m:45s] Cells: 426906                            

 [█████████-------------------------------] 20% [00m:46s] Cells: 431626                            

 [█████████-------------------------------] 20% [00m:47s] Cells: 436346                            

 [█████████-------------------------------] 20% [00m:48s] Cells: 441057                            

 [█████████-------------------------------] 20% [00m:49s] Cells: 445760                            

 [█████████-------------------------------] 20% [00m:50s] Cells: 450532                            

 [█████████-------------------------------] 20% [00m:51s] Cells: 455124                            

 [█████████-------------------------------] 21% [00m:52s] Cells: 459785                            

 [█████████-------------------------------] 22% [00m:53s] Cells: 464457                            

 [█████████-------------------------------] 22% [00m:54s] Cells: 469005                            

 [█████████-------------------------------] 22% [00m:55s] Cells: 473611                            

 [██████████------------------------------] 23% [00m:56s] Cells: 478071                            

 [█████████-------------------------------] 22% [00m:57s] Cells: 482490                            

 [██████████------------------------------] 23% [00m:58s] Cells: 486962                            

 [██████████------------------------------] 23% [00m:59s] Cells: 491391                            

 [██████████------------------------------] 24% [01m:00s] Cells: 495911                            

 [██████████------------------------------] 24% [01m:01s] Cells: 500358                            

 [██████████------------------------------] 24% [01m:02s] Cells: 504644                            

 [██████████------------------------------] 24% [01m:03s] Cells: 508995                            

 [███████████-----------------------------] 25% [01m:04s] Cells: 513295                            

 [███████████-----------------------------] 25% [01m:05s] Cells: 517674                            

 [███████████-----------------------------] 25% [01m:06s] Cells: 521712                            

 [███████████-----------------------------] 25% [01m:07s] Cells: 525848                            

 [██████████------------------------------] 24% [01m:08s] Cells: 530132                            

 [██████████------------------------------] 24% [01m:09s] Cells: 534163                            

 [███████████-----------------------------] 25% [01m:10s] Cells: 538257                            

 [███████████-----------------------------] 25% [01m:11s] Cells: 542351                            

 [██████████------------------------------] 24% [01m:12s] Cells: 546275                            

 [███████████-----------------------------] 25% [01m:13s] Cells: 550387                            

 [███████████-----------------------------] 25% [01m:14s] Cells: 554542                            

 [██████████------------------------------] 24% [01m:15s] Cells: 558496                            

 [██████████------------------------------] 24% [01m:16s] Cells: 562278                            

 [███████████-----------------------------] 25% [01m:17s] Cells: 566174                            

 [███████████-----------------------------] 25% [01m:18s] Cells: 570186                            

 [███████████-----------------------------] 25% [01m:19s] Cells: 574097                            

 [███████████-----------------------------] 25% [01m:20s] Cells: 578042                            

 [███████████-----------------------------] 25% [01m:21s] Cells: 581612                            

 [███████████-----------------------------] 26% [01m:22s] Cells: 585319                            

 [███████████-----------------------------] 26% [01m:23s] Cells: 589190                            

 [███████████-----------------------------] 26% [01m:24s] Cells: 592856                            

 [███████████-----------------------------] 26% [01m:25s] Cells: 596754                            

 [███████████-----------------------------] 27% [01m:26s] Cells: 600813                            

 [███████████-----------------------------] 27% [01m:27s] Cells: 604324                            

 [███████████-----------------------------] 27% [01m:28s] Cells: 607892                            

 [███████████-----------------------------] 27% [01m:29s] Cells: 611596                            

 [███████████-----------------------------] 27% [01m:30s] Cells: 615344                            

 [███████████-----------------------------] 27% [01m:31s] Cells: 619064                            

 [████████████----------------------------] 28% [01m:32s] Cells: 622724                            

 [████████████----------------------------] 28% [01m:33s] Cells: 626400                            

 [████████████----------------------------] 28% [01m:34s] Cells: 629973                            

 [████████████----------------------------] 29% [01m:35s] Cells: 633461                            

 [████████████----------------------------] 29% [01m:36s] Cells: 636999                            

 [████████████----------------------------] 29% [01m:37s] Cells: 640791                            

 [████████████----------------------------] 29% [01m:38s] Cells: 644193                            

 [████████████----------------------------] 29% [01m:39s] Cells: 647751                            

 [█████████████---------------------------] 30% [01m:40s] Cells: 651453                            

 [█████████████---------------------------] 30% [01m:41s] Cells: 654996                            

 [█████████████---------------------------] 30% [01m:42s] Cells: 658463                            

 [█████████████---------------------------] 30% [01m:43s] Cells: 661711                            

 [█████████████---------------------------] 31% [01m:44s] Cells: 665094                            

 [█████████████---------------------------] 31% [01m:45s] Cells: 668695                            

 [█████████████---------------------------] 32% [01m:46s] Cells: 672082                            

 [█████████████---------------------------] 31% [01m:47s] Cells: 675665                            

 [█████████████---------------------------] 31% [01m:48s] Cells: 679028                            

 [█████████████---------------------------] 31% [01m:49s] Cells: 682400                            

 [█████████████---------------------------] 31% [01m:50s] Cells: 685943                            

 [█████████████---------------------------] 31% [01m:51s] Cells: 689415                            

 [█████████████---------------------------] 32% [01m:52s] Cells: 692606                            

 [█████████████---------------------------] 32% [01m:53s] Cells: 695930                            

 [██████████████--------------------------] 33% [01m:54s] Cells: 699184                            

 [██████████████--------------------------] 34% [01m:55s] Cells: 702444                            

 [██████████████--------------------------] 34% [01m:56s] Cells: 705798                            

 [███████████████-------------------------] 35% [01m:57s] Cells: 709239                            

 [███████████████-------------------------] 35% [01m:58s] Cells: 712394                            

 [██████████████--------------------------] 34% [01m:59s] Cells: 715623                            

 [███████████████-------------------------] 35% [02m:00s] Cells: 718932                            

 [███████████████-------------------------] 35% [02m:01s] Cells: 722097                            

 [███████████████-------------------------] 35% [02m:02s] Cells: 725448                            

 [███████████████-------------------------] 36% [02m:03s] Cells: 728520                            

 [███████████████-------------------------] 36% [02m:04s] Cells: 731658                            

 [███████████████-------------------------] 36% [02m:05s] Cells: 734972                            

 [███████████████-------------------------] 36% [02m:06s] Cells: 738330                            

 [███████████████-------------------------] 36% [02m:07s] Cells: 741523                            

 [███████████████-------------------------] 36% [02m:08s] Cells: 744498                            

 [███████████████-------------------------] 36% [02m:09s] Cells: 747856                            

 [███████████████-------------------------] 36% [02m:10s] Cells: 750817                            

 [███████████████-------------------------] 37% [02m:11s] Cells: 754087                            

 [███████████████-------------------------] 37% [02m:12s] Cells: 757193                            

 [███████████████-------------------------] 37% [02m:13s] Cells: 760189                            

 [███████████████-------------------------] 36% [02m:14s] Cells: 763464                            

 [███████████████-------------------------] 36% [02m:15s] Cells: 766385                            

 [███████████████-------------------------] 36% [02m:16s] Cells: 769427                            

 [███████████████-------------------------] 36% [02m:17s] Cells: 772209                            

 [███████████████-------------------------] 36% [02m:18s] Cells: 775077                            

 [███████████████-------------------------] 36% [02m:19s] Cells: 777855                            

 [███████████████-------------------------] 37% [02m:20s] Cells: 780704                            

 [███████████████-------------------------] 37% [02m:21s] Cells: 783439                            

 [███████████████-------------------------] 37% [02m:22s] Cells: 786180                            

 [███████████████-------------------------] 37% [02m:23s] Cells: 788886                            

 [███████████████-------------------------] 37% [02m:24s] Cells: 791646                            

 [███████████████-------------------------] 37% [02m:25s] Cells: 794255                            

 [███████████████-------------------------] 37% [02m:26s] Cells: 796776                            

 [████████████████------------------------] 38% [02m:27s] Cells: 799239                            

 [████████████████------------------------] 38% [02m:28s] Cells: 801743                            

 [████████████████------------------------] 38% [02m:29s] Cells: 803821                            

 [████████████████------------------------] 38% [02m:30s] Cells: 806162                            

 [████████████████------------------------] 38% [02m:31s] Cells: 808037                            

 [████████████████------------------------] 38% [02m:32s] Cells: 810213                            

 [████████████████------------------------] 38% [02m:33s] Cells: 812253                            

 [████████████████------------------------] 38% [02m:34s] Cells: 814507                            

 [████████████████------------------------] 39% [02m:35s] Cells: 816407                            

 [████████████████------------------------] 39% [02m:36s] Cells: 818338                            

 [████████████████------------------------] 39% [02m:37s] Cells: 820206                            

 [████████████████------------------------] 39% [02m:38s] Cells: 822221                            

 [████████████████------------------------] 39% [02m:39s] Cells: 824274                            

 [████████████████------------------------] 39% [02m:40s] Cells: 826212                            

 [████████████████------------------------] 39% [02m:41s] Cells: 827950                            

 [████████████████------------------------] 39% [02m:42s] Cells: 829810                            

 [████████████████------------------------] 39% [02m:43s] Cells: 831652                            

 [█████████████████-----------------------] 40% [02m:44s] Cells: 833475                            

 [█████████████████-----------------------] 41% [02m:45s] Cells: 835211                            

 [█████████████████-----------------------] 41% [02m:46s] Cells: 836990                            

 [█████████████████-----------------------] 41% [02m:47s] Cells: 838644                            

 [█████████████████-----------------------] 41% [02m:48s] Cells: 840471                            

 [█████████████████-----------------------] 40% [02m:49s] Cells: 842178                            

 [█████████████████-----------------------] 40% [02m:50s] Cells: 843958                            

 [█████████████████-----------------------] 41% [02m:51s] Cells: 845560                            

 [█████████████████-----------------------] 41% [02m:52s] Cells: 847238                            

 [█████████████████-----------------------] 41% [02m:53s] Cells: 848800                            

 [█████████████████-----------------------] 42% [02m:54s] Cells: 850234                            

 [█████████████████-----------------------] 42% [02m:55s] Cells: 851960                            

 [█████████████████-----------------------] 42% [02m:56s] Cells: 853720                            

 [██████████████████----------------------] 43% [02m:57s] Cells: 855185                            

 [█████████████████-----------------------] 42% [02m:58s] Cells: 856697                            

 [██████████████████----------------------] 43% [02m:59s] Cells: 858193                            

 [██████████████████----------------------] 43% [03m:00s] Cells: 859706                            

 [██████████████████----------------------] 43% [03m:01s] Cells: 861305                            

 [██████████████████----------------------] 43% [03m:02s] Cells: 862680                            

 [██████████████████----------------------] 44% [03m:03s] Cells: 864214                            

 [██████████████████----------------------] 44% [03m:04s] Cells: 865661                            

 [██████████████████----------------------] 44% [03m:05s] Cells: 867174                            

 [██████████████████----------------------] 44% [03m:06s] Cells: 868581                            

 [███████████████████---------------------] 45% [03m:07s] Cells: 870104                            

 [███████████████████---------------------] 45% [03m:08s] Cells: 871344                            

 [███████████████████---------------------] 45% [03m:09s] Cells: 872641                            

 [███████████████████---------------------] 45% [03m:10s] Cells: 874091                            

 [██████████████████----------------------] 44% [03m:11s] Cells: 875519                            

 [██████████████████----------------------] 44% [03m:12s] Cells: 876636                            

 [██████████████████----------------------] 44% [03m:13s] Cells: 878136                            

 [███████████████████---------------------] 45% [03m:14s] Cells: 879663                            

 [███████████████████---------------------] 45% [03m:15s] Cells: 880911                            

 [███████████████████---------------------] 45% [03m:16s] Cells: 881975                            

 [███████████████████---------------------] 46% [03m:17s] Cells: 883155                            

 [███████████████████---------------------] 46% [03m:18s] Cells: 884517                            

 [███████████████████---------------------] 47% [03m:19s] Cells: 885646                            

 [████████████████████--------------------] 48% [03m:20s] Cells: 886842                            

 [████████████████████--------------------] 48% [03m:21s] Cells: 887670                            

 [████████████████████--------------------] 49% [03m:22s] Cells: 888993                            

 [████████████████████--------------------] 49% [03m:23s] Cells: 890250                            

 [████████████████████--------------------] 48% [03m:24s] Cells: 891359                            

 [████████████████████--------------------] 48% [03m:25s] Cells: 892609                            

 [████████████████████--------------------] 49% [03m:26s] Cells: 893907                            

 [████████████████████--------------------] 49% [03m:27s] Cells: 894895                            

 [████████████████████--------------------] 49% [03m:28s] Cells: 896102                            

 [████████████████████--------------------] 49% [03m:29s] Cells: 897186                            

 [█████████████████████-------------------] 50% [03m:30s] Cells: 898174                            

 [████████████████████--------------------] 49% [03m:31s] Cells: 899331                            

 [████████████████████--------------------] 49% [03m:32s] Cells: 900540                            

 [█████████████████████-------------------] 50% [03m:33s] Cells: 901639                            

 [█████████████████████-------------------] 50% [03m:34s] Cells: 902697                            

 [█████████████████████-------------------] 50% [03m:35s] Cells: 903806                            

 [█████████████████████-------------------] 51% [03m:36s] Cells: 904784                            

 [█████████████████████-------------------] 51% [03m:37s] Cells: 905716                            

 [█████████████████████-------------------] 51% [03m:38s] Cells: 906668                            

 [█████████████████████-------------------] 51% [03m:39s] Cells: 907758                            

 [█████████████████████-------------------] 52% [03m:40s] Cells: 908831                            

 [█████████████████████-------------------] 52% [03m:41s] Cells: 909995                            

 [█████████████████████-------------------] 52% [03m:42s] Cells: 911009                            

 [█████████████████████-------------------] 52% [03m:43s] Cells: 911920                            

 [██████████████████████------------------] 52% [03m:44s] Cells: 912975                            

 [██████████████████████------------------] 52% [03m:45s] Cells: 914026                            

 [██████████████████████------------------] 54% [03m:46s] Cells: 915166                            

 [██████████████████████------------------] 54% [03m:47s] Cells: 916151                            

 [██████████████████████------------------] 54% [03m:48s] Cells: 917167                            

 [██████████████████████------------------] 54% [03m:49s] Cells: 918177                            

 [██████████████████████------------------] 54% [03m:50s] Cells: 918962                            

 [██████████████████████------------------] 54% [03m:51s] Cells: 920092                            

 [██████████████████████------------------] 54% [03m:52s] Cells: 921085                            

 [██████████████████████------------------] 54% [03m:53s] Cells: 921907                            

 [██████████████████████------------------] 54% [03m:54s] Cells: 922865                            

 [███████████████████████-----------------] 55% [03m:55s] Cells: 923866                            

 [██████████████████████------------------] 54% [03m:56s] Cells: 924739                            

 [███████████████████████-----------------] 55% [03m:57s] Cells: 925802                            

 [███████████████████████-----------------] 55% [03m:58s] Cells: 926623                            

 [███████████████████████-----------------] 55% [03m:59s] Cells: 927318                            

 [███████████████████████-----------------] 55% [04m:00s] Cells: 928053                            

 [███████████████████████-----------------] 55% [04m:01s] Cells: 929027                            

 [██████████████████████------------------] 54% [04m:02s] Cells: 929801                            

 [███████████████████████-----------------] 55% [04m:03s] Cells: 930532                            

 [███████████████████████-----------------] 55% [04m:04s] Cells: 931523                            

 [███████████████████████-----------------] 55% [04m:05s] Cells: 932410                            

 [███████████████████████-----------------] 56% [04m:06s] Cells: 933190                            

 [███████████████████████-----------------] 55% [04m:07s] Cells: 934004                            

 [███████████████████████-----------------] 56% [04m:08s] Cells: 934795                            

 [███████████████████████-----------------] 56% [04m:09s] Cells: 935514                            

 [███████████████████████-----------------] 56% [04m:10s] Cells: 936098                            

 [███████████████████████-----------------] 56% [04m:11s] Cells: 937134                            

 [███████████████████████-----------------] 57% [04m:12s] Cells: 937863                            

 [███████████████████████-----------------] 57% [04m:13s] Cells: 938529                            

 [███████████████████████-----------------] 57% [04m:14s] Cells: 939370                            

 [███████████████████████-----------------] 57% [04m:15s] Cells: 940227                            

 [████████████████████████----------------] 58% [04m:16s] Cells: 940761                            

 [████████████████████████----------------] 58% [04m:17s] Cells: 941576                            

 [████████████████████████----------------] 58% [04m:18s] Cells: 942422                            

 [████████████████████████----------------] 58% [04m:19s] Cells: 942884                            

 [████████████████████████----------------] 58% [04m:20s] Cells: 943574                            

 [████████████████████████----------------] 58% [04m:21s] Cells: 944323                            

 [████████████████████████----------------] 58% [04m:22s] Cells: 945216                            

 [████████████████████████----------------] 58% [04m:23s] Cells: 945811                            

 [████████████████████████----------------] 58% [04m:24s] Cells: 946669                            

 [████████████████████████----------------] 58% [04m:25s] Cells: 947380                            

 [████████████████████████----------------] 58% [04m:26s] Cells: 947995                            

 [█████████████████████████---------------] 60% [04m:27s] Cells: 948752                            

 [█████████████████████████---------------] 60% [04m:28s] Cells: 949260                            

 [████████████████████████----------------] 58% [04m:29s] Cells: 950040                            

 [█████████████████████████---------------] 60% [04m:30s] Cells: 950566                            

 [█████████████████████████---------------] 60% [04m:31s] Cells: 951312                            

 [█████████████████████████---------------] 61% [04m:32s] Cells: 952182                            

 [█████████████████████████---------------] 61% [04m:33s] Cells: 952736                            

 [█████████████████████████---------------] 61% [04m:34s] Cells: 953374                            

 [█████████████████████████---------------] 61% [04m:35s] Cells: 953967                            

 [█████████████████████████---------------] 60% [04m:36s] Cells: 954719                            

 [█████████████████████████---------------] 61% [04m:37s] Cells: 955527                            

 [█████████████████████████---------------] 61% [04m:38s] Cells: 956055                            

 [█████████████████████████---------------] 61% [04m:39s] Cells: 956703                            

 [█████████████████████████---------------] 60% [04m:40s] Cells: 957255                            

 [█████████████████████████---------------] 61% [04m:41s] Cells: 957904                            

 [█████████████████████████---------------] 61% [04m:42s] Cells: 958391                            

 [█████████████████████████---------------] 61% [04m:43s] Cells: 959207                            

 [█████████████████████████---------------] 62% [04m:44s] Cells: 959851                            

 [█████████████████████████---------------] 61% [04m:45s] Cells: 960395                            

 [█████████████████████████---------------] 62% [04m:46s] Cells: 961056                            

 [█████████████████████████---------------] 62% [04m:47s] Cells: 961531                            

 [█████████████████████████---------------] 62% [04m:48s] Cells: 962181                            

 [█████████████████████████---------------] 62% [04m:49s] Cells: 962818                            

 [██████████████████████████--------------] 63% [04m:50s] Cells: 963244                            

 [██████████████████████████--------------] 63% [04m:51s] Cells: 963892                            

 [██████████████████████████--------------] 63% [04m:52s] Cells: 964371                            

 [██████████████████████████--------------] 63% [04m:53s] Cells: 964917                            

 [██████████████████████████--------------] 63% [04m:54s] Cells: 965550                            

 [██████████████████████████--------------] 64% [04m:55s] Cells: 965907                            

 [██████████████████████████--------------] 64% [04m:56s] Cells: 966468                            

 [██████████████████████████--------------] 64% [04m:57s] Cells: 967019                            

 [██████████████████████████--------------] 64% [04m:58s] Cells: 967709                            

 [██████████████████████████--------------] 64% [04m:59s] Cells: 968221                            

 [██████████████████████████--------------] 64% [05m:00s] Cells: 968719                            

 [██████████████████████████--------------] 64% [05m:01s] Cells: 969124                            

 [███████████████████████████-------------] 65% [05m:02s] Cells: 969544                            

 [██████████████████████████--------------] 64% [05m:03s] Cells: 970087                            

 [███████████████████████████-------------] 65% [05m:04s] Cells: 970632                            

 [███████████████████████████-------------] 65% [05m:05s] Cells: 971119                            

 [███████████████████████████-------------] 65% [05m:06s] Cells: 971660                            

 [███████████████████████████-------------] 66% [05m:07s] Cells: 971970                            

 [███████████████████████████-------------] 66% [05m:08s] Cells: 972657                            

 [███████████████████████████-------------] 66% [05m:09s] Cells: 972953                            

 [███████████████████████████-------------] 67% [05m:10s] Cells: 973450                            

 [███████████████████████████-------------] 67% [05m:11s] Cells: 973931                            

 [███████████████████████████-------------] 67% [05m:12s] Cells: 974382                            

 [███████████████████████████-------------] 67% [05m:13s] Cells: 974899                            

 [███████████████████████████-------------] 67% [05m:14s] Cells: 975222                            

 [████████████████████████████------------] 68% [05m:15s] Cells: 975672                            

 [████████████████████████████------------] 68% [05m:16s] Cells: 976138                            

 [████████████████████████████------------] 68% [05m:17s] Cells: 976516                            

 [████████████████████████████------------] 68% [05m:18s] Cells: 976635                            

 [████████████████████████████------------] 69% [05m:19s] Cells: 977113                            

 [████████████████████████████------------] 69% [05m:20s] Cells: 977472                            

 [████████████████████████████------------] 69% [05m:21s] Cells: 977938                            

 [█████████████████████████████-----------] 70% [05m:22s] Cells: 978336                            

 [█████████████████████████████-----------] 70% [05m:23s] Cells: 978761                            

 [█████████████████████████████-----------] 70% [05m:24s] Cells: 978929                            

 [█████████████████████████████-----------] 71% [05m:25s] Cells: 979202                            

 [█████████████████████████████-----------] 70% [05m:26s] Cells: 979719                            

 [█████████████████████████████-----------] 71% [05m:27s] Cells: 980104                            

 [█████████████████████████████-----------] 71% [05m:28s] Cells: 980518                            

 [█████████████████████████████-----------] 72% [05m:29s] Cells: 980765                            

 [█████████████████████████████-----------] 72% [05m:30s] Cells: 981106                            

 [█████████████████████████████-----------] 72% [05m:31s] Cells: 981413                            

 [█████████████████████████████-----------] 72% [05m:32s] Cells: 981746                            

 [█████████████████████████████-----------] 72% [05m:33s] Cells: 982264                            

 [█████████████████████████████-----------] 72% [05m:34s] Cells: 982435                            

 [██████████████████████████████----------] 73% [05m:35s] Cells: 982540                            

 [██████████████████████████████----------] 73% [05m:36s] Cells: 983027                            

 [██████████████████████████████----------] 73% [05m:37s] Cells: 983315                            

 [██████████████████████████████----------] 73% [05m:38s] Cells: 983502                            

 [██████████████████████████████----------] 73% [05m:39s] Cells: 983722                            

 [██████████████████████████████----------] 74% [05m:40s] Cells: 984106                            

 [██████████████████████████████----------] 74% [05m:41s] Cells: 984363                            

 [██████████████████████████████----------] 74% [05m:42s] Cells: 984587                            

 [██████████████████████████████----------] 74% [05m:43s] Cells: 984924                            

 [███████████████████████████████---------] 75% [05m:44s] Cells: 985054                            

 [███████████████████████████████---------] 76% [05m:45s] Cells: 985371                            

 [███████████████████████████████---------] 76% [05m:46s] Cells: 985679                            

 [███████████████████████████████---------] 77% [05m:47s] Cells: 985882                            

 [███████████████████████████████---------] 77% [05m:48s] Cells: 986362                            

 [███████████████████████████████---------] 77% [05m:49s] Cells: 986463                            

 [████████████████████████████████--------] 78% [05m:50s] Cells: 986604                            

 [████████████████████████████████--------] 78% [05m:51s] Cells: 986912                            

 [████████████████████████████████--------] 79% [05m:52s] Cells: 987128                            

 [████████████████████████████████--------] 79% [05m:53s] Cells: 987362                            

 [████████████████████████████████--------] 78% [05m:54s] Cells: 987533                            

 [████████████████████████████████--------] 79% [05m:55s] Cells: 987760                            

 [████████████████████████████████--------] 78% [05m:56s] Cells: 987957                            

 [████████████████████████████████--------] 79% [05m:57s] Cells: 988309                            

 [████████████████████████████████--------] 79% [05m:58s] Cells: 988605                            

 [████████████████████████████████--------] 79% [05m:59s] Cells: 988790                            

 [█████████████████████████████████-------] 80% [06m:00s] Cells: 988805                            

 [█████████████████████████████████-------] 81% [06m:01s] Cells: 989121                            

 [█████████████████████████████████-------] 80% [06m:02s] Cells: 989356                            

 [█████████████████████████████████-------] 81% [06m:03s] Cells: 989451                            

 [█████████████████████████████████-------] 81% [06m:04s] Cells: 989625                            

 [█████████████████████████████████-------] 81% [06m:05s] Cells: 989814                            

 [█████████████████████████████████-------] 80% [06m:06s] Cells: 990027                            

 [█████████████████████████████████-------] 80% [06m:07s] Cells: 990236                            

 [█████████████████████████████████-------] 81% [06m:08s] Cells: 990326                            

 [█████████████████████████████████-------] 81% [06m:09s] Cells: 990531                            

 [█████████████████████████████████-------] 81% [06m:10s] Cells: 990732                            

 [█████████████████████████████████-------] 81% [06m:11s] Cells: 990986                            

 [█████████████████████████████████-------] 81% [06m:12s] Cells: 991028                            

 [█████████████████████████████████-------] 82% [06m:13s] Cells: 991055                            

 [█████████████████████████████████-------] 82% [06m:14s] Cells: 991323                            

 [█████████████████████████████████-------] 82% [06m:15s] Cells: 991470                            

 [█████████████████████████████████-------] 82% [06m:16s] Cells: 991559                            

 [█████████████████████████████████-------] 82% [06m:17s] Cells: 991785                            

 [██████████████████████████████████------] 83% [06m:18s] Cells: 991838                            

 [██████████████████████████████████------] 83% [06m:19s] Cells: 991849                            

 [██████████████████████████████████------] 83% [06m:20s] Cells: 992015                            

 [██████████████████████████████████------] 83% [06m:21s] Cells: 992169                            

 [██████████████████████████████████------] 83% [06m:22s] Cells: 992209                            

 [██████████████████████████████████------] 83% [06m:23s] Cells: 992269                            

 [█████████████████████████████████-------] 82% [06m:24s] Cells: 992228                            

 [██████████████████████████████████------] 83% [06m:25s] Cells: 992531                            

 [██████████████████████████████████------] 83% [06m:26s] Cells: 992559                            

 [██████████████████████████████████------] 83% [06m:27s] Cells: 992660                            

 [██████████████████████████████████------] 83% [06m:28s] Cells: 992617                            

 [██████████████████████████████████------] 83% [06m:29s] Cells: 992905                            

 [██████████████████████████████████------] 83% [06m:30s] Cells: 992856                            

 [██████████████████████████████████------] 83% [06m:31s] Cells: 992737                            

 [██████████████████████████████████------] 83% [06m:32s] Cells: 992798                            

 [██████████████████████████████████------] 83% [06m:33s] Cells: 993035                            

 [██████████████████████████████████------] 83% [06m:34s] Cells: 993065                            

 [██████████████████████████████████------] 84% [06m:35s] Cells: 993012                            

 [██████████████████████████████████------] 84% [06m:36s] Cells: 993039                            

 [███████████████████████████████████-----] 85% [06m:37s] Cells: 993053                            

 [███████████████████████████████████-----] 86% [06m:38s] Cells: 993291                            

 [███████████████████████████████████-----] 85% [06m:39s] Cells: 993306                            

 [███████████████████████████████████-----] 86% [06m:40s] Cells: 993164                            

 [███████████████████████████████████-----] 87% [06m:41s] Cells: 993163                            

 [███████████████████████████████████-----] 87% [06m:42s] Cells: 993227                            

 [████████████████████████████████████----] 88% [06m:43s] Cells: 993167                            

 [████████████████████████████████████----] 88% [06m:44s] Cells: 993414                            

 [████████████████████████████████████----] 88% [06m:45s] Cells: 993188                            

 [████████████████████████████████████----] 88% [06m:46s] Cells: 993210                            

 [████████████████████████████████████----] 88% [06m:47s] Cells: 993102                            

 [████████████████████████████████████----] 89% [06m:48s] Cells: 993201                            

 [████████████████████████████████████----] 89% [06m:49s] Cells: 993276                            

 [████████████████████████████████████----] 89% [06m:50s] Cells: 993236                            

 [████████████████████████████████████----] 89% [06m:51s] Cells: 993331                            

 [█████████████████████████████████████---] 90% [06m:52s] Cells: 993189                            

 [█████████████████████████████████████---] 90% [06m:53s] Cells: 993316                            

 [█████████████████████████████████████---] 90% [06m:54s] Cells: 993250                            

 [█████████████████████████████████████---] 91% [06m:55s] Cells: 993192                            

 [█████████████████████████████████████---] 91% [06m:56s] Cells: 993210                            

 [█████████████████████████████████████---] 90% [06m:57s] Cells: 993260                            

 [█████████████████████████████████████---] 91% [06m:58s] Cells: 993203                            

 [█████████████████████████████████████---] 91% [06m:59s] Cells: 993274                            

 [█████████████████████████████████████---] 91% [07m:00s] Cells: 993217                            

 [█████████████████████████████████████---] 91% [07m:01s] Cells: 993198                            

 [█████████████████████████████████████---] 91% [07m:02s] Cells: 993147                            

 [█████████████████████████████████████---] 91% [07m:03s] Cells: 993321                            

 [█████████████████████████████████████---] 91% [07m:04s] Cells: 993429                            

 [█████████████████████████████████████---] 91% [07m:05s] Cells: 993337                            

 [█████████████████████████████████████---] 91% [07m:06s] Cells: 993351                            

 [█████████████████████████████████████---] 91% [07m:07s] Cells: 993135                            

 [█████████████████████████████████████---] 92% [07m:08s] Cells: 993301                            

 [█████████████████████████████████████---] 92% [07m:09s] Cells: 993384                            

 [█████████████████████████████████████---] 92% [07m:10s] Cells: 993280                            

 [█████████████████████████████████████---] 92% [07m:11s] Cells: 993136                            

 [█████████████████████████████████████---] 92% [07m:12s] Cells: 993390                            

 [█████████████████████████████████████---] 92% [07m:13s] Cells: 993208                            

 [██████████████████████████████████████--] 93% [07m:14s] Cells: 993407                            

 [██████████████████████████████████████--] 93% [07m:15s] Cells: 993193                            

 [██████████████████████████████████████--] 94% [07m:16s] Cells: 993246                            

 [██████████████████████████████████████--] 94% [07m:17s] Cells: 993371                            

 [███████████████████████████████████████-] 95% [07m:18s] Cells: 993151                            

 [██████████████████████████████████████--] 94% [07m:19s] Cells: 993257                            

 [███████████████████████████████████████-] 95% [07m:20s] Cells: 993288                            

 [██████████████████████████████████████--] 94% [07m:21s] Cells: 993204                            

 [██████████████████████████████████████--] 94% [07m:22s] Cells: 993408                            

 [██████████████████████████████████████--] 94% [07m:23s] Cells: 993372                            

 [██████████████████████████████████████--] 94% [07m:24s] Cells: 993215                            

 [███████████████████████████████████████-] 95% [07m:25s] Cells: 993208                            

 [███████████████████████████████████████-] 95% [07m:26s] Cells: 993255                            

 [███████████████████████████████████████-] 95% [07m:27s] Cells: 993233                            

 [███████████████████████████████████████-] 95% [07m:28s] Cells: 993330                            

 [███████████████████████████████████████-] 95% [07m:29s] Cells: 993192                            

 [███████████████████████████████████████-] 95% [07m:30s] Cells: 993121                            

 [███████████████████████████████████████-] 96% [07m:31s] Cells: 993247                            

 [███████████████████████████████████████-] 97% [07m:32s] Cells: 993296                            

 [███████████████████████████████████████-] 97% [07m:33s] Cells: 993228                            

 [███████████████████████████████████████-] 97% [07m:34s] Cells: 993202                            

 [███████████████████████████████████████-] 97% [07m:35s] Cells: 993275                            

 [████████████████████████████████████████] 98% [07m:36s] Cells: 993204                            

 [████████████████████████████████████████] 98% [07m:37s] Cells: 993232                            

 [███████████████████████████████████████-] 97% [07m:38s] Cells: 993297                            

 [███████████████████████████████████████-] 97% [07m:39s] Cells: 993258                            

 [███████████████████████████████████████-] 96% [07m:40s] Cells: 993423                            

 [███████████████████████████████████████-] 97% [07m:41s] Cells: 993168                            

 [████████████████████████████████████████] 98% [07m:42s] Cells: 993221                            

 [████████████████████████████████████████] 98% [07m:43s] Cells: 993318                            

 [████████████████████████████████████████] 99% [07m:44s] Cells: 993211                            

 [████████████████████████████████████████] 99% [07m:45s] Cells: 993186                            

 [████████████████████████████████████████] 100% [07m:45s] Saving snapshot                         

# find a 50x50 sample containing 80 "B" cells and 10 "A" cells at least
sim$search_sample(c("A" = 10, "B" = 80), 50, 50)
#> TissueRectangle((450,450),(499,499))