Simulating the cell evolution in a tissue

This class simulates the cell evolution on a tissue.

Details

The objects of this class can simulate the evolution of many cells belonging to different species on a tissue. Each cell can duplicate or die according to the rates that delineate the cell species.

TissueSimulation() supports epigenetic evolutions, and it lets users define species pairs that belong to the same mutant (even though, its genomic characterisation is unknown) and differ because of their epigenetic states.

TissueSimulation() models epigenetic mutations and allows a cell in one of mutant species to generate a new cell belonging to the other species of the same mutant at a specified rate.

TissueSimulation() also allows users to schedule mutations from one mutant to a different mutant.

This class provides the following methods and properties:

add_epigenetic_state() adds an epigenetic state and its species.

add_epigenetic_states() adds epigenetic states and their species.

add_mutant() adds a mutant and its species.

add_mutants() adds mutants and their species.

border_growth_model is a Boolean flag to enable/disable the border-driven growth model.

choose_border_cell_in() randomly chooses a cell in tumour border.

choose_cell_in() randomly chooses a tumour cell.

death_activation_level stores the minimum number of cells in a species to enable death.

get_added_cells() returns the cells manually added to the simulation.

get_cell() returns one of the simulation cells by its position.

get_cells() returns some of the simulation cells by their positions.

get_clock() returns the current simulation time.

get_count_history() returns the number of cells by species along the simulation.

get_counts() returns the number of cells by species.

get_epigenetic_states() returns the names of the simulated epigenetic states.

get_firing_history() returns the number of events fired along the simulation.

get_firings() returns the number of fired events.

get_lineage_graph() returns the lineage graph.

get_mutants() returns the names of the simulated mutants.

get_name() returns the name of the simulation.

get_rates() returns the tissue simulation rates.

get_rates_update_history() returns the history of the rates along the simulation.

get_sample_forest() returns the sample forest of the tissue simulation.

get_samples_info() returns information about the collected samples.

get_snapshot_info() returns information about the collected snapshots.

get_tissue_size() returns the size of the simulated tissue.

history_delta stores the time difference between two consecutive samples of the tissue simulation status.

mutate_progeny() duplicates a cell and mutates one of its children.

place_cell() places a cell in the simulated tissue.

run_until() simulates the evolution of the tissue until a condition holds.

run_up_to_event() simulates the evolution of the tissue until the number of events of either a mutant or a species is below a specified threshold.

run_up_to_size() simulates the evolution of the tissue until the number of cells of either a mutant or a species is below a specified threshold .

run_up_to_time() simulates the evolution of the tissue until the specified simulation time.

sample_cells() samples tissue cells.

schedule_mutation() schedules a mutation.

search_sample() searches for a rectangular tissue sample satisfying some conditions.

search_samples() searches for rectangular tissue samples satisfying some conditions.

set_rate() sets the rate of an event.

set_rates() sets the rates of multiple events.

snapshot_trigger exposes the tissue simulation snapshot triggers.

var() builds a simulation variable.

See also